requirements.txt
django 
python-decouple 
psycopg2-binary
django-trix-fork
pillow
channels
channels-redis
django-extensions